+++ src/Connection/Driver/MariaDBCapabilities.php
@@ -59,4 +59,12 @@ class MariaDBCapabilities implements DriverCapabilities
     {
         return true;
     }
+
+    /**
+     * @inheritDoc
+     */
+    public function getIdentifierQuote(): string
+    {
+        return '`';
+    }
 }