+++ src/Connection/Driver/PostgreSQLCapabilities.php
@@ -59,4 +59,12 @@ class PostgreSQLCapabilities implements DriverCapabilities
     {
         return true;
     }
+
+    /**
+     * @inheritDoc
+     */
+    public function getIdentifierQuote(): string
+    {
+        return '"';
+    }
 }